Combining postInChat and isUnkenny in shared.js file.
