--- src/scripts/unkenny.js
@@ -1,9 +1,0 @@
-function isUnkenny(actor) {
-    if (!actor) {
-        return false;
-    }
-    let preamble = actor.getFlag("unkenny", "preamble");
-    return !!preamble;
-}
-
-export { isUnkenny };
